+++ packages/mcp-server/src/tools/blocklist/retrieve-blocklist.ts
@@ -1,5 +1,6 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
+import { maybeFilter } from 'hyperswitch-mcp/filtering';
 import { asTextContentResult } from 'hyperswitch-mcp/tools/types';
 
 import { Tool } from '@modelcontextprotocol/sdk/types.js';
@@ -17,12 +18,19 @@ export const metadata: Metadata = {
 
 export const tool: Tool = {
   name: 'retrieve_blocklist',
-  description: '',
+  description:
+    "When using this tool, always use the `jq_filter` parameter to reduce the response size and improve performance.\n\nOnly omit if you're sure you don't need the data.\n\n\n\n# Response Schema\n```json\n{\n  $ref: '#/$defs/blocklist_response',\n  $defs: {\n    blocklist_response: {\n      type: 'object',\n      properties: {\n        created_at: {\n          type: 'string',\n          format: 'date-time'\n        },\n        data_kind: {\n          $ref: '#/$defs/blocklist_data_kind'\n        },\n        fingerprint_id: {\n          type: 'string'\n        }\n      },\n      required: [        'created_at',\n        'data_kind',\n        'fingerprint_id'\n      ]\n    },\n    blocklist_data_kind: {\n      type: 'string',\n      enum: [        'payment_method',\n        'card_bin',\n        'extended_card_bin'\n      ]\n    }\n  }\n}\n```",
   inputSchema: {
     type: 'object',
     properties: {
       data_kind: {
         $ref: '#/$defs/blocklist_data_kind',
+      },
+      jq_filter: {
+        type: 'string',
+        title: 'jq Filter',
+        description:
+          'A jq filter to apply to the response to include certain fields. Consult the output schema in the tool description to see the fields that are available.\n\nFor example: to include only the `name` field in every object of a results array, you can provide ".results[].name".\n\nFor more information, see the [jq documentation](https://jqlang.org/manual/).',
       },
     },
     $defs: {
@@ -36,7 +44,7 @@ export const tool: Tool = {
 
 export const handler = async (client: Hyperswitch, args: Record<string, unknown> | undefined) => {
   const body = args as any;
-  return asTextContentResult(await client.blocklist.retrieve(body));
+  return asTextContentResult(await maybeFilter(args, await client.blocklist.retrieve(body)));
 };
 
 export default { metadata, tool, handler };