+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.0.1-alpha.2 (2025-07-11)
+
+Full Changelog: [v0.0.1-alpha.1...v0.0.1-alpha.2](https://github.com/gorakhnathy7/hs-mcp/compare/v0.0.1-alpha.1...v0.0.1-alpha.2)
+
+### Chores
+
+* add docs to RequestOptions type ([b19a4e9](https://github.com/gorakhnathy7/hs-mcp/commit/b19a4e9c549f751fe6c53f15a7888a530bb0fa56))
+* **internal:** codegen related update ([1eba97c](https://github.com/gorakhnathy7/hs-mcp/commit/1eba97c82c9501f1be5d958502e9d069bb99cc0a))
+* make some internal functions async ([a4cd427](https://github.com/gorakhnathy7/hs-mcp/commit/a4cd42713bbe37d5c2b1441359676a5cb2dd50c2))
+
 ## 0.0.1-alpha.1 (2025-07-01)
 
 Full Changelog: [v0.0.1-alpha.0...v0.0.1-alpha.1](https://github.com/gorakhnathy7/hs-mcp/compare/v0.0.1-alpha.0...v0.0.1-alpha.1)